separate function for session credentials
